\name{mlThurstone-package}
\alias{mlThurstone-package}
\alias{mlThurstone}
\docType{package}
\title{
\packageTitle{mlThurstone}
}
\description{
\packageDescription{mlThurstone}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mlThurstone}
\packageIndices{mlThurstone}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{mlThurstone}

Maintainer: \packageMaintainer{mlThurstone}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
